requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements/dev.in
#
asgiref==3.3.1
    # via django
certifi==2020.12.5
    # via requests
chardet==4.0.0
    # via requests
click==7.1.2
    # via pip-tools
coreapi==2.3.3
    # via drf-yasg
coreschema==0.0.4
    # via
    #   coreapi
    #   drf-yasg
django-choices==1.7.1
    # via -r requirements/base.in
django-widget-tweaks==1.4.8
    # via -r requirements/base.in
django==3.1.5
    # via
    #   -r requirements/base.in
    #   django-choices
    #   djangorestframework
    #   drf-yasg
djangorestframework==3.12.2
    # via
    #   -r requirements/base.in
    #   drf-yasg
drf-yasg==1.20.0
    # via -r requirements/base.in
idna==2.10
    # via requests
inflection==0.5.1
    # via drf-yasg
itypes==1.2.0
    # via coreapi
jinja2==2.11.3
    # via coreschema
markupsafe==1.1.1
    # via jinja2
packaging==20.9
    # via drf-yasg
pip-tools==5.5.0
    # via -r requirements/dev.in
pyparsing==2.4.7
    # via packaging
pytz==2021.1
    # via django
requests==2.25.1
    # via coreapi
ruamel.yaml.clib==0.2.2
    # via ruamel.yaml
ruamel.yaml==0.16.12
    # via drf-yasg
six==1.15.0
    # via django-choices
sqlparse==0.4.1
    # via django
uritemplate==3.0.1
    # via
    #   coreapi
    #   drf-yasg
urllib3==1.26.3
    # via requests

# The following packages are considered to be unsafe in a requirements file:
# pip
